# 89% Caching improvement - Async Redis completion calls + batch redis GET requests for a given key + call type 