--- a/litellm/proxy/_new_secret_config.yaml
+++ b/litellm/proxy/_new_secret_config.yaml
@@ -13,7 +13,7 @@ litellm_settings:
   cache: true
   cache_params:
     type: redis
-  # callbacks: ["batch_redis_requests"]
+  callbacks: ["batch_redis_requests"]
 
 general_settings:
   master_key: sk-1234