- 89% improvement - Moving all our requests to async, improvement proxy perf w/ caching from 9419.55ms to 979.79ms. 
- 7.2% improvement - New (optional) Batch Redis hook, further improves this from 979.79ms to 909.07ms

## Batch Redis 

**What it does?**
When a request is made:
- Check if a key starting with `litellm:<hashed_api_key>:<call_type>:` exists in-memory, if no - get the last 100 cached requests for this key and store it 
- New requests are stored with this `litellm:..` as the namespace 

**Why?**
Reduce number of redis GET requests. 

**Usage**
```yaml
litellm_settings:
  cache: true
  cache_params:
    type: redis
  callbacks: ["batch_redis_requests"]
```